fix: limitTIme 타입 변경

+++ src/main/kotlin/com/leita/leita/domain/problem/Limit.kt
@@ -9,5 +9,5 @@ open class Limit(
     val memory: Long,
 
     @Column(name = "limit_time")
-    val time: Double,
+    val time: Long,
 )
